Move the view logic into a separate file
